# app/swagger/swagger.yaml
basePath: /
definitions:
  config.DevopsSysConfig:
    properties:
      createId:
        description: 创建者
        type: integer
      created_at:
        type: string
      id:
        description: 主键
        type: integer
      key:
        description: 配置的key
        type: string
      name:
        description: 配置名
        type: string
      remark:
        description: 描述
        type: string
      updateId:
        description: 更新人
        type: integer
      updated_at:
        type: string
      value:
        description: 配置的value
        type: string
    type: object
  domain.DevopsSysDomain:
    properties:
      createId:
        description: 创建者
        type: integer
      created_at:
        type: string
      domainNum:
        description: 商户号
        type: string
      enable:
        description: 是否启用
        type: integer
      englishName:
        description: 英文名
        type: string
      id:
        description: 主键
        type: integer
      name:
        description: 域账户名
        type: string
      remark:
        description: 描述
        type: string
      updateId:
        description: 更新人
        type: integer
      updated_at:
        type: string
    type: object
...
  path.DevopsSysApi:
    properties:
      apiGroup:
        description: api组
        type: string
      createId:
        description: 创建者
        type: integer
      created_at:
        type: string
      description:
        description: api中文描述
        type: string
      id:
        description: 主键
        type: integer
      method:
        description: 方法:创建POST(默认)|查看GET|更新PUT|删除DELETE
        type: string
      path:
        description: api路径
        type: string
      updateId:
        description: 更新人
        type: integer
      updated_at:
        type: string
    type: object
...
  request.PageRequest:
    properties:
      filter:
        items:
          type: string
        type: array
      page:
        type: integer
      pageSize:
        type: integer
    type: object
...
  request.SearchApiParams:
    properties:
      apiGroup:
        description: api组
        type: string
      createId:
        description: 创建者
        type: integer
      created_at:
        type: string
      desc:
        description: 排序方式:升序false(默认)|降序true
        type: boolean
      description:
        description: api中文描述
        type: string
      filter:
        items:
          type: string
        type: array
      id:
        description: 主键
        type: integer
      method:
        description: 方法:创建POST(默认)|查看GET|更新PUT|删除DELETE
        type: string
      orderKey:
        description: 排序
        type: string
      page:
        type: integer
      pageSize:
        type: integer
      path:
        description: api路径
        type: string
      updateId:
        description: 更新人
        type: integer
      updated_at:
        type: string
    type: object
  request.SearchConfigParams:
    properties:
      createId:
        description: 创建者
        type: integer
      created_at:
        type: string
      desc:
        description: 排序方式:升序false(默认)|降序true
        type: boolean
      filter:
        items:
          type: string
        type: array
      id:
        description: 主键
        type: integer
      key:
        description: 配置的key
        type: string
      name:
        description: 配置名
        type: string
      orderKey:
        description: 排序
        type: string
      page:
        type: integer
      pageSize:
        type: integer
      remark:
        description: 描述
        type: string
      updateId:
        description: 更新人
        type: integer
      updated_at:
        type: string
      value:
        description: 配置的value
        type: string
    type: object
...
  response.PageResult:
    properties:
      columns: {}
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  sys.RequestById:
    properties:
      ids:
        description: 角色ID
        type: string
    type: object
...
info:
  contact: {}
  description: This is Devops API
  title: Swagger Devops API
  version: 0.0.1
paths:
  /sys/api/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: api路径, api中文描述, api组, 方法
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/path.DevopsSysApi'
      produces:
      - application/json
      responses:
        "200":
          description: 创建基础api
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建基础api
      tags:
      - Apis
...
  /sys/api/delete:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/path.DevopsSysApi'
      produces:
      - application/json
      responses:
        "200":
          description: 删除api
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除api
      tags:
      - Apis
...
  /sys/api/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 分页获取API列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.SearchApiParams'
      produces:
      - application/json
      responses:
        "200":
          description: 分页获取API列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取API列表
      tags:
      - Apis
  /sys/api/modify:
    put:
      consumes:
      - application/json
      parameters:
      - description: api路径, api中文描述, api组, 方法
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/path.DevopsSysApi'
      produces:
      - application/json
      responses:
        "200":
          description: 修改基础api
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 修改基础api
      tags:
      - Apis
...
  /sys/config/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: api路径, api中文描述, api组, 方法
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/config.DevopsSysConfig'
      produces:
      - application/json
      responses:
        "200":
          description: 创建基础api
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建config
      tags:
      - Config
  /sys/config/delete:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/sys.RequestById'
      produces:
      - application/json
      responses:
        "200":
          description: 删除api
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除config
      tags:
      - Config
  /sys/config/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 分页获取API列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.SearchConfigParams'
      produces:
      - application/json
      responses:
        "200":
          description: 分页获取API列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取config列表
      tags:
      - Config
  /sys/config/modify:
    put:
      consumes:
      - application/json
      parameters:
      - description: api路径, api中文描述, api组, 方法
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/config.DevopsSysConfig'
      produces:
      - application/json
      responses:
        "200":
          description: 修改基础api
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 修改config
      tags:
      - Config
  /sys/domain/{id}:
    get:
      consumes:
      - application/json
      description: 获得域接口
      parameters:
      - description: 域id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获得域接口
      tags:
      - Domain
  /sys/domain/add:
    post:
      consumes:
      - application/json
      description: 新增域接口
      parameters:
      - description: 域
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/domain.DevopsSysDomain'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 新增域接口
      tags:
      - Domain
  /sys/domain/delete:
    delete:
      consumes:
      - application/json
      description: 删除域接口
      parameters:
      - description: 域ids
        in: body
        name: ids
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除域接口
      tags:
      - Domain
  /sys/domain/list:
    post:
      consumes:
      - application/json
      description: 获得域列表接口
      parameters:
      - description: 页数，页大小，筛选条件
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.PageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.DevopsSysDomain'
      security:
      - ApiKeyAuth: []
      summary: 获得域列表接口
      tags:
      - Domain
  /sys/domain/modify:
    put:
      consumes:
      - application/json
      description: 修改域接口
      parameters:
      - description: 域
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/domain.DevopsSysDomain'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 修改域接口
      tags:
      - Domain
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: bearer
    type: apiKey
swagger: "2.0"